requirements.txt
Django==1.7.5
https://github.com/whalesalad/django-restless/archive/master.zip
Fabric==1.9.0
Jinja2==2.7.2
MarkupSafe==0.18
PyJWT==0.1.8
PyYAML==3.10
Pygments==1.6
astroid==1.0.1
backports.ssl-match-hostname==3.4.0.2
boto==2.23.0
distribute==0.7.3
django-appconf==0.6
https://github.com/django-extensions/django-extensions/archive/master.zip
django-hstore==1.2.1
django-redis==3.5.2
djorm-pgarray==1.0.2
ecdsa==0.10
futures==2.1.6
gevent==1.0
greenlet==0.4.2
hiredis==0.1.2
kafka-python==0.9.2
logilab-common==0.61.0
mock==1.0.1
paramiko==1.12.1
psycopg2==2.5.2
pycrypto==2.6.1
redgreenunittest==0.1.0
redis==2.9.1
requests==2.2.1
six==1.5.2
tornado==3.2
ujson==1.33
wsgiref==0.1.2
yolk==0.4.3
testscenarios==0.4
netaddr==0.7.11
celery==3.1.15
flower==0.6.0
lxml==3.3.5
pytz==2014.2
coverage==3.7.1
colorlog==2.0.0
python-statsd==1.6.3
python-logstash==0.3.1
pyOpenSSL==0.14
dictdiffer==0.1.0
arrow==0.4.4
kairos==0.10.1
py2neo==2.0.5